Feat: UrlTransformer를 인터페이스 형식으로 변경하고 BASE62 기반 구현체를 추가한다.
